+++ src/ska_mid_dish_manager/component_managers/dish_manager_cm.py
@@ -4,7 +4,7 @@ import logging
 import os
 from functools import partial
 from threading import Lock
-from typing import Callable, Optional, Tuple
+from typing import Callable, List, Optional, Tuple
 
 import tango
 from ska_control_model import CommunicationStatus, HealthState, ResultCode, TaskStatus
@@ -222,7 +222,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
             ],
         }
 
-    def _get_active_sub_component_managers(self) -> dict:
+    def _get_active_sub_component_managers(self) -> List:
         """Get a list of subservient device component managers which are not being ignored."""
         active_component_managers = [self.sub_component_managers["DS"]]
 
@@ -741,20 +741,30 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
         self,
         task_callback: Optional[Callable] = None,
     ) -> Tuple[TaskStatus, str]:
-        """Transition the dish to STOW mode"""
-        _is_set_stow_mode_allowed = partial(
-            self._dish_mode_model.is_command_allowed,
-            "SetStowMode",
-            component_manager=self,
-            task_callback=task_callback,
+        """Transition the dish to STOW mode
+
+        Note: To expedite the command, it does not
+        implement _is_track_stow_cmd_allowed() because by
+        default it is allowed to run at all states.
+        """
+        ds_cm = self.sub_component_managers["DS"]
+        try:
+            ds_cm.execute_command("Stow", None)
+
+        except (LostConnection, tango.DevFailed) as err:
+            task_callback(status=TaskStatus.FAILED, exception=err)
+            return TaskStatus.FAILED, f"{err}"
+        task_callback(
+            progress="Stow called, monitor dishmode for LRC completed", status=TaskStatus.COMPLETED
         )
-        status, response = self.submit_task(
-            self._command_map.set_stow_mode,
-            args=[],
-            is_cmd_allowed=_is_set_stow_mode_allowed,
-            task_callback=task_callback,
-        )
-        return status, response
+        # abort queued tasks on the task executor's threadpoolexecutor
+        self.abort_commands()
+        # abort the task on the subservient devices
+        sub_component_mgrs = self._get_active_sub_component_managers()
+        for component_mgr in sub_component_mgrs:
+            component_mgr.abort_commands()
+
+        return TaskStatus.COMPLETED, "Stow called, monitor dishmode for LRC completed"
 
     def slew(
         self,